requirements.txt
# Django Backend
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# PDF Processing
PyPDF2>=3.0.0

# Machine Learning & Text Processing
scikit-learn>=1.3.0
numpy>=1.24.0

# Streamlit Frontend
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0